refactor: move UserAvatar components in accordance with atomic design
