+++ src/components/Text.tsx
@@ -1,6 +1,6 @@
 import { format, parseISO } from 'date-fns';
 import type TextType from '../lib/models/Text';
-import UserAvatar from './UserAvatar';
+import UserAvatar from './atoms/UserAvatar';
 import UserLink from './UserLink';
 import UserName from './UserName';
