+++ src/pages/users/[id].tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import Container from '../../components/atoms/Container';
 import NavBar from '../../components/molecules/NavBar';
 import TextList from '../../components/TextList';
-import UserAvatar from '../../components/UserAvatar';
+import UserAvatar from '../../components/atoms/UserAvatar';
 import UserName from '../../components/UserName';
 import useUser from '../../hooks/useUser';
